+++ src/internal/api_info/collect_info_test.go
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-package restlerwrapper
+package api_info
