+++ src/internal/model/container_image.go
