+++ src/internal/persistence/redis/container_image_test.go
